Separate tours redux into separate files
