+++ app/javascript/packs/application/redux/tours/selectors.js
@@ -1,0 +1,7 @@
+export const getTours = state => {
+  return Object.values(state.tours.data);
+};
+
+export const isFetching = state => {
+  return state.tours.isFetching;
+};
